requirements.txt
Pillow==10.3.0
piexif==1.1.3